complete move copy link story
